move athena and statsd also
