+++ release/files_common
@@ -77,12 +77,12 @@ system/tombstoned.py
 selfdrive/pandad.py
 selfdrive/updated.py
 system/rtshield.py
-selfdrive/statsd.py
+system/statsd.py
 
-selfdrive/athena/__init__.py
-selfdrive/athena/athenad.py
-selfdrive/athena/manage_athenad.py
-selfdrive/athena/registration.py
+system/athena/__init__.py
+system/athena/athenad.py
+system/athena/manage_athenad.py
+system/athena/registration.py
 
 selfdrive/boardd/.gitignore
 selfdrive/boardd/SConscript
