+++ selfdrive/thermald/power_monitoring.py
@@ -9,7 +9,7 @@ from common.params import Params, put_nonblocking
 from common.realtime import sec_since_boot
 from system.hardware import HARDWARE
 from system.swaglog import cloudlog
-from selfdrive.statsd import statlog
+from system.statsd import statlog
 
 CAR_VOLTAGE_LOW_PASS_K = 0.091 # LPF gain for 5s tau (dt/tau / (dt/tau + 1))
